#54 refactor: 팀 컨벤션 맞춰 수정
