+++ src/features/auth/kakao/model/useKakaoLogin.ts
@@ -1,16 +1,15 @@
 import {useNavigate} from 'react-router-dom';
 import {useUserStore} from '@/entities/auth/model/useUserStore';
-import {useKakaoMutation} from '@/features/auth/kakao/api/useKakaoMutation';
+import {authMutations} from '@/entities/auth/api/authMutations';
+import {useMutation} from '@tanstack/react-query';
 import {useCallback} from 'react';
 
-/**
- * 카카오 로그인 프로세스를 처리하는 커스텀 훅
- */
 export const useKakaoLogin = () => {
   const navigate = useNavigate();
   const {login} = useUserStore();
 
-  const {mutate} = useKakaoMutation({
+  const {mutate} = useMutation({
+    ...authMutations.kakaoLogin,
     onSuccess: ({response}) => {
       const userType = response.role === 'ADMIN' ? 'admin' : 'student';
       login(response.name, userType, response.accessToken);
